Share the archive and file helpers both installers need
- Move the zip media type check into ModArchive, so the mod installer and the loader installer accept the same archives.
- Carve the temporary-file-and-move write out of TomlFileStore into AtomicFile, so a configuration file can be written the same way as a TOML file.
- Word the escaping-entry error for any install folder, not only a mod folder.
- Share the fake downloader and the zip builder between the installer tests.
